drivers: camera: Add switch for enable/disable boot clock timestamp
Our kernel currently supports both AOSP and MIUI ROMs, however
the current AOSP ROM does not include the stock libcameraservice
blob and it uses the timestamp buffer of the system monotonic
time. If using the timestamp buffer of the boot clock this
results in duration issue when recording video, so let's add
switch to enable/disable the timestamp buffer of boot time.

Signed-off-by: Fiqri Ardyansyah <fiqri0927936@gmail.com>