ion: Add support for heap info queries
On Android 12, libdmabufheap requires non-legacy ION implementations to
support heap info queries [1]. Lack of support for the ioctl causes the
assertion to fail, which breaks camera.

Implement the ioctl by porting Google's implementation and adjusting the
ioctl path accordingly to support different output formats. This also
requires us to track the total heap count as userspace uses it to
calculate the size of the result buffer.

[1] https://android.googlesource.com/platform/system/memory/libdmabufheap/+/f2c9b847808f1528c2ab735db64ae61525a813c6/BufferAllocator.cpp#83

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Fiqri Ardyansyah <fiqri0927936@gmail.com>